@@ -8,10 +8,10 @@ readme = "README.md"
 packages = [{ include = "nonebot_plugin_saa" }]
 classifiers = [
   "Development Status :: 3 - Alpha",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
   "Operating System :: Microsoft :: Windows",
   "Operating System :: POSIX :: Linux",
   "Operating System :: MacOS",
@@ -19,8 +19,8 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.8"
-nonebot2 = "^2.0.0"
+python = ">=3.9, <4.0"
+nonebot2 = "^2.3.0"
 strenum = "^0.4.8"
 pydantic = ">=1.10.0,<3.0.0,!=2.5.0,!=2.5.1"
 anyio = ">=3.3.0,<5.0.0"
@@ -30,7 +30,7 @@ filetype = "^1.2.0"
 pytest-asyncio = "^0.20.2"
 ipdb = "^0.13.9"
 pip = "^23.0"
-pytest-cov = "^4.0.0"
+pytest-cov = "^5.0.0"
 nonebug = "^0.3.1"
 nonebot2 = { extras = [
   "fastapi",
@@ -56,7 +56,7 @@ nonebot-adapter-satori = "^0.10.2"
 
 [tool.black]
 line-length = 88
-target-version = ["py38", "py39", "py310"]
+target-version = ["py39", "py310", "py311", "py312"]
 include = '\.pyi?$'
 extend-exclude = '''
 '''
@@ -67,6 +67,7 @@ line_length = 88
 length_sort = true
 skip_gitignore = true
 force_sort_within_sections = true
+src_paths = ["nonebot_plugin_saa", "tests"]
 extra_standard_library = ["typing_extensions"]
 
 [build-system]
@@ -77,9 +78,43 @@ build-backend = "poetry.core.masonry.api"
 asyncio_mode = "auto"
 
 [tool.pyright]
-pythonVersion = "3.8"
+pythonVersion = "3.9"
 pythonPlatform = "All"
 defineConstant = { PYDANTIC_V2 = true }
+typeCheckingMode = "basic"
+reportShadowedImports = false
+disableBytesTypePromotions = true
+
+[tool.ruff]
+line-length = 88
+target-version = "py39"
+
+[tool.ruff.lint]
+select = [
+  "F",     # Pyflakes
+  "W",     # pycodestyle warnings
+  "E",     # pycodestyle errors
+  "UP",    # pyupgrade
+  "ASYNC", # flake8-async
+  "C4",    # flake8-comprehensions
+  "T10",   # flake8-debugger
+  "T20",   # flake8-print
+  "PYI",   # flake8-pyi
+  "PT",    # flake8-pytest-style
+  "Q",     # flake8-quotes
+  "RUF",   # Ruff-specific rules
+]
+ignore = [
+  "E402",   # module-import-not-at-top-of-file
+  "UP037",  # quoted-annotation
+  "RUF001", # ambiguous-unicode-character-string
+  "RUF002", # ambiguous-unicode-character-docstring
+  "RUF003", # ambiguous-unicode-character-comment
+]
+
+[tool.ruff.lint.flake8-pytest-style]
+fixture-parentheses = false
+mark-parentheses = false
 
 [tool.coverage.report]
 exclude_also = [
