@@ -1,15 +1,15 @@
 from io import BytesIO
 from pathlib import Path
+from typing import Literal
 from functools import partial
-from typing import List, Literal
 
 from nonebot.adapters import Event
 from nonebot.compat import model_dump
 from nonebot.adapters import Bot as BaseBot
 
 from ..auto_select_bot import register_list_targets
-from ..utils import SupportedAdapters, SupportedPlatform
 from ..types import Text, Image, Reply, Mention, MentionAll
+from ..utils import SupportedAdapters, SupportedPlatform, type_message_id_check
 from ..abstract_factories import (
     MessageFactory,
     register_ms_adapter,
@@ -101,7 +101,7 @@ try:
 
     @register_onebot_v12(Reply)
     async def _reply(r: Reply) -> MessageSegment:
-        assert isinstance(mid := r.data["message_id"], OB12MessageId)
+        mid = type_message_id_check(OB12MessageId, r.data["message_id"])
         return MessageSegment.reply(mid.message_id)
 
     @register_target_extractor(PrivateMessageEvent)
@@ -314,7 +314,7 @@ try:
         return OB12Receipt(bot_id=bot.self_id, message_id=message_id)
 
     @register_list_targets(SupportedAdapters.onebot_v12)
-    async def list_targets(bot: BaseBot) -> List[PlatformTarget]:
+    async def list_targets(bot: BaseBot) -> list[PlatformTarget]:
         assert isinstance(bot, Bot)
 
         targets = []
