@@ -1,17 +1,14 @@
 import inspect
-from typing_extensions import Annotated
+from collections.abc import Awaitable
 from typing import (
     TYPE_CHECKING,
     Any,
-    Dict,
-    Type,
-    Tuple,
     Union,
     Literal,
     Callable,
     ClassVar,
     Optional,
-    Awaitable,
+    Annotated,
     cast,
 )
 
@@ -92,9 +89,9 @@ class TargetQQGroupOpenId(PlatformTarget, BotSpecifier):
         group_openid: 群 open_id
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.qq_group_openid] = (
         SupportedPlatform.qq_group_openid
-    ] = SupportedPlatform.qq_group_openid
+    )
     group_openid: str
 
 
@@ -105,9 +102,9 @@ class TargetQQPrivateOpenId(PlatformTarget, BotSpecifier):
         user_openid: 用户 open_id
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.qq_private_openid] = (
         SupportedPlatform.qq_private_openid
-    ] = SupportedPlatform.qq_private_openid
+    )
     user_openid: str
 
 
@@ -118,9 +115,9 @@ class TargetQQGuildChannel(PlatformTarget):
         channel_id: 子频道号
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.qq_guild_channel] = (
         SupportedPlatform.qq_guild_channel
-    ] = SupportedPlatform.qq_guild_channel
+    )
     channel_id: int
 
 
@@ -132,9 +129,9 @@ class TargetQQGuildDirect(PlatformTarget):
         source_guild_id: 来自的频道号
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.qq_guild_direct] = (
         SupportedPlatform.qq_guild_direct
-    ] = SupportedPlatform.qq_guild_direct
+    )
     recipient_id: int
     source_guild_id: int
 
@@ -147,9 +144,9 @@ class TargetOB12Unknow(PlatformTarget):
         user_id, group_id, channel_id, guild_id: 同 ob12 定义
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.unknown_ob12] = (
         SupportedPlatform.unknown_ob12
-    ] = SupportedPlatform.unknown_ob12
+    )
     platform: str
     detail_type: Literal["private", "group", "channel"]
     user_id: Optional[str] = None
@@ -168,9 +165,9 @@ class TargetSatoriUnknown(PlatformTarget):
         channel_id: 频道 ID
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.unknown_satori] = (
         SupportedPlatform.unknown_satori
-    ] = SupportedPlatform.unknown_satori
+    )
     platform: str
     user_id: Optional[str] = None
     guild_id: Optional[str] = None
@@ -184,9 +181,9 @@ class TargetKaiheilaChannel(PlatformTarget):
         channel_id: 频道ID
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.kaiheila_channel] = (
         SupportedPlatform.kaiheila_channel
-    ] = SupportedPlatform.kaiheila_channel
+    )
     channel_id: str
 
 
@@ -197,9 +194,9 @@ class TargetKaiheilaPrivate(PlatformTarget):
         user_id: 接收人ID
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.kaiheila_private] = (
         SupportedPlatform.kaiheila_private
-    ] = SupportedPlatform.kaiheila_private
+    )
     user_id: str
 
 
@@ -210,9 +207,9 @@ class TargetTelegramCommon(PlatformTarget):
         user_id: 对话ID
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.telegram_common] = (
         SupportedPlatform.telegram_common
-    ] = SupportedPlatform.telegram_common
+    )
     chat_id: Union[int, str]
 
 
@@ -224,9 +221,9 @@ class TargetTelegramForum(PlatformTarget):
         message_thread_id: 板块ID
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.telegram_forum] = (
         SupportedPlatform.telegram_forum
-    ] = SupportedPlatform.telegram_forum
+    )
     chat_id: int
     message_thread_id: int
 
@@ -238,9 +235,9 @@ class TargetFeishuPrivate(PlatformTarget):
         open_id: 用户 Open ID
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.feishu_private] = (
         SupportedPlatform.feishu_private
-    ] = SupportedPlatform.feishu_private
+    )
     open_id: str
 
 
@@ -251,9 +248,9 @@ class TargetFeishuGroup(PlatformTarget):
         chat_id: 群 ID
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.feishu_group] = (
         SupportedPlatform.feishu_group
-    ] = SupportedPlatform.feishu_group
+    )
     chat_id: str
 
 
@@ -265,9 +262,9 @@ class TargetDoDoChannel(PlatformTarget):
         dodo_source_id: 用户 ID(可选)
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.dodo_channel] = (
         SupportedPlatform.dodo_channel
-    ] = SupportedPlatform.dodo_channel
+    )
     channel_id: str
     dodo_source_id: Optional[str] = None
 
@@ -280,9 +277,9 @@ class TargetDoDoPrivate(PlatformTarget):
         island_source_id: 群 ID
     """
 
-    platform_type: Literal[
+    platform_type: Literal[SupportedPlatform.dodo_private] = (
         SupportedPlatform.dodo_private
-    ] = SupportedPlatform.dodo_private
+    )
     island_source_id: str
     dodo_source_id: str
 
@@ -308,8 +305,8 @@ AllSupportedPlatformTarget = Union[
 ]
 
 
-ConvertToArg = Callable[[PlatformTarget], Dict[str, Any]]
-convert_to_arg_map: Dict[Tuple[SupportedPlatform, SupportedAdapters], ConvertToArg] = {}
+ConvertToArg = Callable[[PlatformTarget], dict[str, Any]]
+convert_to_arg_map: dict[tuple[SupportedPlatform, SupportedAdapters], ConvertToArg] = {}
 
 
 def register_convert_to_arg(adapter: SupportedAdapters, platform: SupportedPlatform):
@@ -322,10 +319,10 @@ def register_convert_to_arg(adapter: SupportedAdapters, platform: SupportedPlatf
 
 Extractor = Callable[[Event], PlatformTarget]
 ExtractorWithBotSpecifier = Callable[[Event, Bot], PlatformTarget]
-extractor_map: Dict[Type[Event], Union[Extractor, ExtractorWithBotSpecifier]] = {}
+extractor_map: dict[type[Event], Union[Extractor, ExtractorWithBotSpecifier]] = {}
 
 
-def register_target_extractor(event: Type[Event]):
+def register_target_extractor(event: type[Event]):
     def wrapper(func: Union[Extractor, ExtractorWithBotSpecifier]):
         extractor_map[event] = func
         return func
@@ -364,7 +361,7 @@ Sender = Callable[
     Awaitable["Receipt"],
 ]
 
-sender_map: Dict[SupportedAdapters, Sender] = {}
+sender_map: dict[SupportedAdapters, Sender] = {}
 
 
 def register_sender(adapter: SupportedAdapters):
@@ -378,7 +375,7 @@ def register_sender(adapter: SupportedAdapters):
 SaaTarget = Annotated[PlatformTarget, Depends(get_target)]
 
 QQGuild_DMS = Callable[[TargetQQGuildDirect, Bot], Awaitable[int]]
-qqguild_dms_map: Dict[SupportedAdapters, QQGuild_DMS] = {}
+qqguild_dms_map: dict[SupportedAdapters, QQGuild_DMS] = {}
 
 
 def register_qqguild_dms(adapter: SupportedAdapters):
@@ -390,7 +387,7 @@ def register_qqguild_dms(adapter: SupportedAdapters):
 
 
 class QQGuildDMSManager:
-    _cache: ClassVar[Dict[TargetQQGuildDirect, int]] = {}
+    _cache: ClassVar[dict[TargetQQGuildDirect, int]] = {}
 
     @classmethod
     def get_guild_id(cls, target: TargetQQGuildDirect) -> int:
